separate classes and modularize generate_reaction()
